Set actin length via n_pairs or length spec
Set the actin length either by directly specifying the number of g-actin
pairs in the filament or by setting the length of the filament and then
back-calculating it from the length. This comes along with a change in
the order of arguments to an actin __init__ function, placing tract
before either of the two length commands.