Prevent diffusion out of tract
Don’t allow freely diffusing proteins (actin and alpha-actinic when
unbound) to diffuse out the ends of the tract. This assumes that the
tract is “blocked” at those ends b/c they are located on the cell
membrane etc.

To avoid those ends acting as “absorbing” attractors, we coerce the
molecules back into the tract by reflecting them across the boundary
condition.

## Deeper background 
Before this there were no boundaries to prevent the diffusion of molecules out of the ends of the tracts. We need to prevent that as we lose unbound molecules fairly quickly. This is possibly due to overestimating the diffusion of components within the fiber. (So we might be able to better address that by taking a longer and more motivated look at the reasoning behind the current level of diffusion expected for unbound components. But, as that is a longer process, we'll confine ourselves to creating confinement for the time being.)