Prevent diffusion out of tract
Don’t allow freely diffusing proteins (actin and alpha-actinic when
unbound) to diffuse out the ends of the tract. This assumes that the
tract is “blocked” at those ends b/c they are located on the cell
membrane etc.

To avoid those ends acting as “absorbing” attractors, we coerce the
molecules back into the tract by reflecting them across the boundary
condition.