Check actin bounds when selecting the nearest site
Previously we had just been calculating the nearest of all binding
sites, checking the distance to each one. We don't need to do this when
we are looking for the nearest site to a location that is past either
end of the filament (we just return the first or last site in those
cases). This is a speed up (depending on the distributions of querying
molecules, the length of the tract, and the length of the actin) of ~2x
for this function, which is a substantial (~20%) fraction of the current
run time.