Actinin uses support.kinetics for rate-to-prob
Move to using the better documented function in the kinetics support module for converting rates to probabilities of transition. Additionally rename transition r12 to r01 and r21 to r10 for consistency with future motors that will have more than two states and which will have data structures that benefit from being zero indexed.