Update actinin location from bound sites when not freely diffusing
This addresses a bug wherein underlying actinin locations were only being set during diffusion. This meant that when an actinin latched on and was dragged away from its initial position the perceived location of the unbound head would slowly diverge from its real value. By updating the underlying actinin location from that of the bound heads we eliminate this drift.